--- server/__init__.py
@@ -1,2 +1,0 @@
-"""Competition Server for Reinforcement Agents -- Teamprojekt WS 23/24.
-"""
