--- teamprojekt_competition_server/__init__.py
@@ -1,4 +1,0 @@
-"""Competition Server for Reinforcement Agents -- Teamprojekt WS 23/24.
-
-TODO: Add a description of the package here.
-"""
