+++ teamprojekt_competition_server/server/__init__.py
@@ -1,0 +1,1 @@
+
